Refactoring EditTimeEntryView to more consistent name scheme.
